agent/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
logfire
httpx
google-generativeai
Pillow